+++ scripts/functions.sh
@@ -17,7 +17,7 @@ error() {
 function display_login_instructions {
     echo ""
     echo "To login the rover to azure:"
-    echo " rover login [subscription_id_to_target(optional)]"
+    echo " rover login [tenant_name.onmicrosoft.com or tenant_guid (optional)] [subscription_id_to_target(optional)]"
     echo ""
     echo " rover logout"
     echo ""
@@ -80,17 +80,17 @@ function verify_azure_session {
         echo "Checking existing Azure session"
         session=$(az account show)
 
-        if [ "${tf_command}" != "login" ] && [ ! -z "${tf_action}" ]; then
-            echo "Login to azure with tenant ${tf_command}"
-            ret=$(az login --tenant ${tf_command} >/dev/null >&1)
+        if [ ! -z "${tf_action}" ]; then
+            echo "Login to azure with tenant ${tf_action}"
+            ret=$(az login --tenant ${tf_action} >/dev/null >&1)
         else
             ret=$(az login >/dev/null >&1)
         fi
 
         # the second parameter would be the subscription id to target
-        if [ ! -z "${tf_action}" ]; then
-            echo "Set default subscription to ${tf_action}"
-            az account set -s ${tf_action}
+        if [ "${tf_command}" != "login" ] && [ ! -z "${tf_command}" ]; then
+            echo "Set default subscription to ${tf_command}"
+            az account set -s ${tf_command}
         fi
         
         az account show
@@ -159,7 +159,7 @@ function initialize_state {
             plan
             apply
             # Create sandpit workspace
-            id=$(az storage account list --query "[?tags.workspace=='level0']" | jq -r .[0].id)
+            id=$(az storage account list --query "[?tags.tfstate=='level0']" | jq -r .[0].id)
             workspace_create "sandpit"
             workspace_create ${TF_VAR_workspace}
             upload_tfstate
@@ -284,7 +284,7 @@ function get_remote_state_details {
     export ARM_ACCESS_KEY=$(az storage account keys list --account-name ${TF_VAR_lowerlevel_storage_account_name} --resource-group ${TF_VAR_lowerlevel_resource_group_name} | jq -r .[0].value)
 
     # Set the security context under the devops app
-    export keyvault=$(az keyvault list --query "[?tags.workspace=='level0']" | jq -r .[0].name) && echo " - keyvault_name: ${keyvault}"
+    export keyvault=$(az keyvault list --query "[?tags.tfstate=='level0']" | jq -r .[0].name) && echo " - keyvault_name: ${keyvault}"
     export TF_VAR_lowerlevel_container_name=$(az keyvault secret show -n launchpad-blob-container --vault-name ${keyvault} | jq -r .value) && echo " - container: ${TF_VAR_lowerlevel_container_name}"
     export TF_VAR_lowerlevel_key=$(az keyvault secret show -n launchpad-blob-name --vault-name ${keyvault} | jq -r .value) && echo " - tfstate file: ${TF_VAR_lowerlevel_key}"